requirements.txt
djangorestframework
markdown
django-filter
django-cron
pandas